requirement.txt
Django==2.1
django-cors-headers==2.2.0
django-extra-fields==1.0.0
django-filter==1.1.0
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.5
